+++ web_app/lexi_vault.py
@@ -1,16 +1,35 @@
 from flask import Flask, render_template, request, jsonify
 from flask_sqlalchemy import SQLAlchemy 
 from sqlalchemy import create_engine
-from models import dictionary
-from sqlalchemy.orm import sessionmaker
+
+# Replace 'username', 'password', 'hostname', 'port', and 'database_name' with your actual credentials
+username = 'lexi_vault'
+password = 'lexipass'
+hostname = 'localhost'  # or your MySQL server's IP address
+port = '3306'  # or your MySQL server's port
+database_name = 'lexi_vault_db'
+
+# Construct the MySQL connection URL
+url = f"mysql+pymysql://{username}:{password}@{hostname}:{port}/{database_name}"
+
+# Create the SQLAlchemy engine
+engine = create_engine(url)
+
+# Now you can use this engine object to interact with your MySQL database
 
 
+app = Flask(__name__)
+#app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite3'
+app.config['SQLALCHEMY_DATABASE_URI'] = url
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
+db = SQLAlchemy(app)
 
-@app.route('/', strict_slashes=False)
-def homepage():
-    return render_template('index.html')
-
+class dictionary(db.Model):
+    word_id = db.Column(db.Integer, primary_key=True)
+    words = db.Column(db.String(80), unique=False, nullable=False)
+    part_of_speech = db.Column(db.String(80), unique=False, nullable=False)
+    meanings = db.Column(db.String(800), unique=False, nullable=False)
 
 @app.route('/lexi_vault', strict_slashes=False, methods=['GET', 'POST'])
 def lexi_vault():
@@ -27,7 +46,7 @@ def lexi_vault():
                 part_of_speech = dictionary_data.part_of_speech
             else:
                 meaning = 'Word Not Found'
-                part_of_speech = 'Please click the link below to add it to list of suggested words'
+                part_of_speech = 'Please add to list of suggested words'
             # all_words_data = dictionary.query.with_entities(dictionary.words).all()
             # all_words = '\n'.join(word[0] for word in dictionary_data)
             return render_template('lexi_vault.html', word=word, meaning=meaning, part_of_speech=part_of_speech, all_words=all_words)
@@ -48,22 +67,6 @@ def get_word_details():
             part_of_speech = dictionary_data.part_of_speech
             return jsonify({'meaning': meaning, 'part_of_speech': part_of_speech})
     return jsonify({'error': 'Word not found'})
-
-
-@app.route('/suggest', strict_slashes=False, methods=['GET', 'POST'])
-def suggest():
-    return render_template('suggest.html')
-
-
-@app.route('/developers', strict_slashes=False)
-def developers():
-    return render_template('developers.html')
-
-
-@app.route('/about', strict_slashes=False)
-def about():
-    return render_template('about.html')
-
