+++ web_app/static/styles/developers.css
@@ -11,7 +11,7 @@ body {
     padding: 20;
     text-align: center;
     padding-top: 10px;
-    background: linear-gradient(to left, rgba(6, 104, 6, 0.89), rgba(18, 158, 158, 0.808), rgb(104, 95, 104));
+    background-color: greenyellow;
 }
 
 .container {
