+++ web_app/templates/index.html
@@ -18,10 +18,10 @@
             <span></span>
             
             <ul class="menu_links">
-                <a href="lexi_vault"><li>Lexi_Vault</li></a>
-                <a href="developers"><li>Developers</li></a>
-                <a href="suggest"><li>Suggest Words</li></a>
-                <a href="about"><li>About</li></a>
+                <a href="lexi_vault.html"><li>Lexi_Vault</li></a>
+                <a href="developers.html"><li>Developers</li></a>
+                <a href="suggest.html"><li>Suggest Words</li></a>
+                <a href="about.html"><li>About</li></a>
             </ul>
         </div>
     </nav>
