+++ web_app/templates/lexi_vault.html
@@ -26,6 +26,5 @@
     <textarea id="allWords" name="wordtext" readonly>{{ all_words }}</textarea>
     <textarea id="searchResult" readonly>{{ word }} - {{ meaning }} - {{ part_of_speech }}</textarea>
 </body>
-
 <a href= "suggest.html"> click here to suggest words to be added to lexi_Vault</a>
 </html>
